+++ packages/form-designer/index.ts
@@ -1,7 +1,7 @@
 export { FormDesigner } from './designer/FormDesigner'
 export type { FormDesignerProps } from './designer/FormDesigner'
 export { getComponent, getMenus, registerComponent } from './registry/registry'
-export type { ComponentDef, ConfigMeta, MenuGroup } from './registry/registry'
+export type { ComponentDef, ConfigMeta, ListRenderCtx, MenuGroup } from './registry/registry'
 export { FormRenderer } from './renderer/FormRenderer'
 export type { FormRendererProps } from './renderer/FormRenderer'
 export type { FieldSchema, FormGlobalConfig, FormSchema, ValidateRule } from './types/schema'